Made xaxis and yaxis explicit options
If we want to have the ROI on something other than just navigation
dimension, we need to be able to specify the axes. This commit adds
explicit xaxis and yaxis, but they still default to navigation_axes 0
and 1 to be backwards compatible (and for ease of use). Should probably
default to signal_axes 0, 1 if not navigation dimensions are correct,
but better to force explicit behavior.

Also fixed some small bugs (None checks etc).